refactor: Improved error handling and increased code security

+++ crud/delete.py
@@ -4,14 +4,20 @@ from colorama import init, Fore, Style
 init()
 
 def deleteCustomers():
-  while True:
-    print(Fore.GREEN + "\nDeletar Clientes" + Style.RESET_ALL)
-    id_customer = int(input("Digite o ID do cliente que você deseja deletar: "))
-    delete_customer(id_customer)
-    
-    print(Fore.GREEN + "\nCliente deletado com sucesso!" + Style.RESET_ALL)
-    break
-    
-    
+    try:
+        while True:
+            id_customer = int(input("\nDigite o ID do cliente que você deseja deletar: "))
+            
+            confirmacao = input(Fore.YELLOW + f"Tem certeza que deseja deletar o cliente com ID {id_customer}? (s/n): " + Style.RESET_ALL).lower()
+            if confirmacao.lower() == 's':
+                delete_customer(id_customer)
+            else:
+                print(Fore.BLUE + "\nAção cancelada." + Style.RESET_ALL)
+            break
+    except ValueError:
+        print(Fore.RED + "\nErro: O ID deve ser um número inteiro válido." + Style.RESET_ALL)
+    except Exception as e:
+        print(Fore.RED + f"\nErro ao deletar o cliente: {str(e)}" + Style.RESET_ALL)
+
 if __name__ == "__main__":
-  deleteCustomers()
+    deleteCustomers()
